chore(invoice): DRY invoice index in controller and test
